pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "notarb-cli"
version = "0.1.0"
description = "CLI monitoring tool for NotArb MEV bot"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "rich>=13.0.0",
    "psutil>=5.9.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/notarb-cli"
Repository = "https://github.com/yourusername/notarb-cli.git"

[project.scripts]
notarb-cli = "notarb_cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["manager*"]